Move contract cleanup code to test utils
